+++ src/com/towson/wavyleaf/Help.java
@@ -1,7 +1,9 @@
 package com.towson.wavyleaf;
 
+import android.content.Context;
 import android.content.Intent;
 import android.content.res.Resources;
+import android.graphics.Bitmap;
 import android.os.Bundle;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
@@ -12,6 +14,7 @@ import com.actionbarsherlock.view.MenuItem;
 public class Help extends SherlockActivity {
 	
 	protected WebView wv;
+	protected Context ctx = this;
 
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
@@ -29,13 +32,30 @@ public class Help extends SherlockActivity {
 		wv.getSettings().setBuiltInZoomControls(true);
 		
 		wv.setWebViewClient(new WebViewClient() {
+			
         	@Override
-        	public void onReceivedError(WebView view, int errorCode, String description, String failingUrl) {}
+        	public void onReceivedError(WebView view, int errorCode, String description, String failingUrl) {
+        		
+        	}
+        	
         	public boolean shouldOverrideUrlLoading(WebView v, String url) {
         		v.loadUrl(url);
         		return true;
         	}
-        });
+
+        	@Override
+			public void onPageStarted(WebView view, String url, Bitmap favicon) {
+				super.onPageStarted(view, url, favicon);
+				setTitle("Loading...");
+			}
+        	
+        	@Override
+			public void onPageFinished(WebView view, String url) {
+				super.onPageFinished(view, url);
+				setTitle("Help");
+			}
+        
+		});
 		
 		Resources res = this.getResources();
         wv.loadUrl(res.getString(R.string.website));