+++ src/com/towson/wavyleaf/Settings.java
@@ -16,7 +16,7 @@ public class Settings extends SherlockPreferenceActivity implements OnSharedPref
 	public static final String KEY_CHECKBOX_NOISE = "preference_noise";
 	public static final String KEY_EMAIL = "preference_email";
 	public static final String KEY_NAME = "preference_name";
-	public static final String KEY_USERNAME = "preference_username";
+	public static final String KEY_USER_ID = "preference_user_id";
 	public static final String KEY_SINGLETALLY = "preference_singletally";
 	public static final String KEY_TRIPTALLY = "preference_triptally";
 	public static final String KEY_TRIPTALLY_CURRENT = "preference_triptally_current"; // Key for tally for only current trip
@@ -76,7 +76,6 @@ public class Settings extends SherlockPreferenceActivity implements OnSharedPref
 		// Instantiation
 		Preference p_age = (Preference) findPreference(KEY_BIRTHYEAR);
 		Preference p_name = (Preference) findPreference(KEY_NAME);
-		Preference p_username = findPreference(KEY_USERNAME);
 		Preference p_tally_single = findPreference(KEY_SINGLETALLY);
 		Preference p_tally_trip = findPreference(KEY_TRIPTALLY);
 		CheckBoxPreference cbp_vibrate = (CheckBoxPreference) findPreference(KEY_CHECKBOX_VIBRATE);
@@ -84,7 +83,6 @@ public class Settings extends SherlockPreferenceActivity implements OnSharedPref
 		
 		// Read values
 		String string_name = sp.getString(KEY_NAME, "null");
-		String string_username = sp.getString(KEY_USERNAME, "null");
 		String string_age = sp.getString(KEY_BIRTHYEAR, "null");
 		int int_tally_single = sp.getInt(KEY_SINGLETALLY, 0);
 		int int_tally_trip = sp.getInt(KEY_TRIPTALLY, 0);
@@ -93,7 +91,6 @@ public class Settings extends SherlockPreferenceActivity implements OnSharedPref
 		
 		// Set Summaries
 		p_name.setSummary(capitalizeFirstLetter(string_name));
-		p_username.setSummary(capitalizeFirstLetter(string_username));
 		p_age.setSummary(capitalizeFirstLetter(string_age));
 		p_tally_single.setSummary(int_tally_single + "");
 		p_tally_trip.setSummary(int_tally_trip + "");